Updates test files to reflect test contents
